暂存

+++ controllers/announcementController.js
@@ -50,9 +50,9 @@ exports.getAnnouncementDetail = async (req, res, next) => {
 
 
 /**
- * @description 新增公告接口（暂未实现权限校验）
+ * @description 新增公告接口
  * @param {Object} req - 请求对象
- * @param {Object} req.user.position - 用户职位，用于校验修改权限       //暂未实现
+ * @param {Object} req.user.groups - 用户组别，用于校验修改权限
  * @param {Object} req.body - 公告数据
  * @param {string} req.body.title - 公告标题 (必填)
  * @param {string} [req.body.cover_url] - 公告封面URL (可选)
@@ -63,10 +63,10 @@ exports.getAnnouncementDetail = async (req, res, next) => {
  */
 exports.createAnnouncement = async (req, res, next) => {
     try {
-        // // 权限校验
-        // if (req.user.position !== 'admin') {
-        //     throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
-        // }
+        // 权限校验
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-president' || g === 'gdmu/NA-president')) {
+            throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
+        }
         const announcementData = req.body;
         const result = await announcementService.createAnnouncement(announcementData);
         return res.success(result, '公告新增成功', 'ANNOUNCEMENT_CREATED');
@@ -77,18 +77,19 @@ exports.createAnnouncement = async (req, res, next) => {
 
 
 /**
- * @description 删除公告接口（暂未实现权限校验）
+ * @description 删除公告接口
  * @param {Object} req - 请求对象  
- * @param {Object} req.user.position - 用户职位，用于校验修改权限       //暂未实现
+ * @param {Object} req.user.groups - 用户组别，用于校验修改权限
  * @param {number} req.params.id - 公告ID
  * @returns {Promise<Object>} 删除结果
  */
 exports.deleteAnnouncement = async (req, res, next) => {
     try {
-        // // 权限校验
-        // if (req.user.position !== 'admin') {
-        //     throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
-        // }
+        console.log(req.user);
+        // 权限校验
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-president' || g === 'gdmu/NA-president')) {
+            throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
+        }
         const announcementId = req.params.id;
         const result = await announcementService.deleteAnnouncement(announcementId);
         return res.success(result, '公告删除成功', 'ANNOUNCEMENT_DELETED');
@@ -99,7 +100,7 @@ exports.deleteAnnouncement = async (req, res, next) => {
 
 
 /**
- * @description 修改公告接口（暂未实现权限校验）
+ * @description 修改公告接口
  * @param {Object} req - 请求对象
  * @param {number} req.user.position - 用户职位，用于校验修改权限       //暂未实现
  * @param {number} req.params.id - 公告ID
@@ -113,10 +114,10 @@ exports.deleteAnnouncement = async (req, res, next) => {
  */
 exports.updateAnnouncement = async (req, res, next) => {
     try {
-        // // 权限校验
-        // if (req.user.position !== 'admin') {
-        //     throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
-        // }
+        // 权限校验
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-president' || g === 'gdmu/NA-president')) {
+            throw new AppError('您没有权限修改该公告', 403, 'NO_PERMISSION');
+        }
         const announcementId = req.params.id;
         const updateData = req.body;
         const result = await announcementService.updateAnnouncement(announcementId, updateData);
